pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud_cost_env"
version = "1.0.0"
description = "Cloud Infrastructure Cost Optimizer for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0"
]